+++ tests/maps/spreadOperatorAllowed.js
@@ -1,6 +1,0 @@
-// testing options.spreadOperatorAllowed
-
-module.exports = [
-
-]
-