+++ tools/build.js
@@ -3,11 +3,12 @@ const fs = require("fs");
 const path = require("path");
 const rimraf = require("rimraf");
 
-let dev = process.env.NODE_ENV !== "production";
-
 let pegjsOptions = {
   output: "source",
   format: "commonjs",
+  features: {
+    expected: false
+  },
 };
 
 let replacements = [